Rename file and use default Overriden Hero
