+++ packages/core/src/components/sections/Hero/OverriddenDefaultHero.ts
